requirements.txt
Django>=4.2.0
djangorestframework>=3.15.0
psycopg2-binary>=2.9.10         # PostgreSQL adapter
drf-yasg>=1.21.8                # Swagger documentation
django-cors-headers>=4.3.1      # CORS for frontend access
gunicorn>=21.2.0                # For production servers
python-decouple>=3.8            # To load .env variables
django-filter>=24.2             # filtering support
django-redis>=5.4.0             # caching (for high traffic)
redis>=4.6.0
pytest-django>=4.5.2
black==24.3.0
flake8>=6.0.0

